--- server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/tracing/TraceJdbcWriter.java
@@ -1,380 +1,0 @@
-/*
- *    Copyright 2020 bithon.org
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.bithon.server.storage.jdbc.tracing;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.extern.slf4j.Slf4j;
-import org.bithon.component.commons.tracing.SpanKind;
-import org.bithon.component.commons.utils.CollectionUtils;
-import org.bithon.component.commons.utils.RetryUtils;
-import org.bithon.component.commons.utils.StringUtils;
-import org.bithon.server.storage.jdbc.common.jooq.Tables;
-import org.bithon.server.storage.tracing.ITraceWriter;
-import org.bithon.server.storage.tracing.TraceSpan;
-import org.bithon.server.storage.tracing.TraceStorageConfig;
-import org.bithon.server.storage.tracing.index.TagIndex;
-import org.bithon.server.storage.tracing.mapping.TraceIdMapping;
-import org.jooq.ConnectionRunnable;
-import org.jooq.DSLContext;
-import org.jooq.Table;
-import org.jooq.TransactionalRunnable;
-import org.jooq.exception.DataAccessException;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.Timestamp;
-import java.time.Duration;
-import java.time.LocalDateTime;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-/**
- * @author frank.chen021@outlook.com
- * @date 30/12/20
- */
-@Slf4j
-public class TraceJdbcWriter implements ITraceWriter {
-
-    public static String getInsertSpanTableStatement(DSLContext dslContext, Table<?> table) {
-        return dslContext.render(dslContext.insertInto(table,
-                                                       Tables.BITHON_TRACE_SPAN.TIMESTAMP,
-                                                       Tables.BITHON_TRACE_SPAN.APPNAME,
-                                                       Tables.BITHON_TRACE_SPAN.INSTANCENAME,
-                                                       Tables.BITHON_TRACE_SPAN.TRACEID,
-                                                       Tables.BITHON_TRACE_SPAN.SPANID,
-                                                       Tables.BITHON_TRACE_SPAN.PARENTSPANID,
-                                                       Tables.BITHON_TRACE_SPAN.NAME,
-                                                       Tables.BITHON_TRACE_SPAN.CLAZZ,
-                                                       Tables.BITHON_TRACE_SPAN.METHOD,
-                                                       Tables.BITHON_TRACE_SPAN.KIND,
-                                                       Tables.BITHON_TRACE_SPAN.STARTTIMEUS,
-                                                       Tables.BITHON_TRACE_SPAN.ENDTIMEUS,
-                                                       Tables.BITHON_TRACE_SPAN.COSTTIMEMS,
-                                                       Tables.BITHON_TRACE_SPAN.ATTRIBUTES,
-                                                       Tables.BITHON_TRACE_SPAN.NORMALIZEDURL,
-                                                       Tables.BITHON_TRACE_SPAN.STATUS)
-                                           .values((LocalDateTime) null,
-                                                   //app name
-                                                   null,
-                                                   // instance
-                                                   null,
-                                                   //trace id
-                                                   null,
-                                                   // span id
-                                                   null,
-                                                   // parent id
-                                                   null,
-                                                   // name
-                                                   null,
-                                                   // class
-                                                   null,
-                                                   // method
-                                                   null,
-                                                   // kind
-                                                   null,
-                                                   // start time
-                                                   null,
-                                                   // end time
-                                                   null,
-                                                   // cost time
-                                                   null,
-                                                   // tags
-                                                   null,
-                                                   // normalized url
-                                                   null,
-                                                   // status
-                                                   null
-                                                  ));
-    }
-
-    protected final DSLContext dslContext;
-    private final TraceStorageConfig traceStorageConfig;
-
-    public TraceJdbcWriter(DSLContext dslContext, TraceStorageConfig traceStorageConfig) {
-        this.dslContext = dslContext;
-        this.traceStorageConfig = traceStorageConfig;
-    }
-
-    protected boolean isTransactionSupported() {
-        return true;
-    }
-
-    public abstract static class InsertSpanRunnable implements IInsert {
-        private final String insertStatement;
-        private final List<TraceSpan> spans;
-        private final String table;
-
-        public InsertSpanRunnable(DSLContext dslContext,
-                                  Table<?> table,
-                                  List<TraceSpan> spans) {
-            this.insertStatement = getInsertSpanTableStatement(dslContext, table);
-            this.table = table.getName();
-            this.spans = spans;
-        }
-
-        @Override
-        public String getTable() {
-            return table;
-        }
-
-        @Override
-        public int getInsertSize() {
-            return spans.size();
-        }
-
-        @Override
-        public void run(Connection connection) throws Exception {
-            try (PreparedStatement statement = connection.prepareStatement(insertStatement)) {
-                for (TraceSpan span : spans) {
-                    statement.setObject(1, new Timestamp(span.startTime / 1000).toLocalDateTime());
-                    statement.setObject(2, span.appName);
-                    statement.setObject(3, span.instanceName);
-                    statement.setObject(4, span.traceId);
-                    statement.setObject(5, span.spanId);
-                    statement.setObject(6, span.parentSpanId);
-                    statement.setObject(7, StringUtils.getOrEmpty(span.name));
-                    statement.setObject(8, StringUtils.getOrEmpty(span.clazz));
-                    statement.setObject(9, StringUtils.getOrEmpty(span.method));
-                    statement.setObject(10, StringUtils.getOrEmpty(span.kind));
-                    statement.setObject(11, span.startTime);
-                    statement.setObject(12, span.endTime);
-                    statement.setObject(13, span.costTime);
-                    statement.setObject(14, toTagStore(span.getTags()));
-                    statement.setObject(15, span.getNormalizedUri());
-                    statement.setObject(16, span.getStatus());
-                    statement.addBatch();
-                }
-
-                RetryUtils.retry(statement::executeBatch,
-                                 this::isExceptionRetryable,
-                                 3,
-                                 Duration.ofMillis(100));
-            }
-        }
-
-        protected abstract Object toTagStore(Map<String, String> tag);
-
-        protected boolean isExceptionRetryable(Exception e) {
-            return false;
-        }
-    }
-
-    private void writeSpans(List<TraceSpan> traceSpans, Table<?> table) throws Throwable {
-        if (!traceSpans.isEmpty()) {
-            doInsert(createInsertSpanRunnable(table, traceSpans));
-        }
-    }
-
-    protected void doInsert(IInsert runnable) throws Throwable {
-        try {
-            dslContext.connection(runnable);
-        } catch (DataAccessException e) {
-            // Re-throw the caused exception for more clear stack trace
-            // In such a case, the caused exception is not NULL.
-            throw e.getCause();
-        }
-    }
-
-    private void writeMappings(List<TraceIdMapping> mappings) {
-        if (CollectionUtils.isEmpty(mappings)) {
-            return;
-        }
-
-        dslContext.connection(createInsertMappingRunnable(dslContext, mappings));
-    }
-
-    public static class InsertMappingRunnable implements ConnectionRunnable {
-        private final String insertStatement;
-        private final Collection<TraceIdMapping> mappings;
-
-        public InsertMappingRunnable(DSLContext dslContext,
-                                     Collection<TraceIdMapping> mappings) {
-            insertStatement = dslContext.render(dslContext.insertInto(Tables.BITHON_TRACE_MAPPING,
-                                                                      Tables.BITHON_TRACE_MAPPING.TRACE_ID,
-                                                                      Tables.BITHON_TRACE_MAPPING.USER_TX_ID,
-                                                                      Tables.BITHON_TRACE_SPAN.TIMESTAMP)
-                                                          .values((String) null, null, null));
-            this.mappings = mappings;
-        }
-
-        @Override
-        public void run(Connection connection) throws Exception {
-            try (PreparedStatement statement = connection.prepareStatement(insertStatement)) {
-                for (TraceIdMapping mapping : mappings) {
-                    statement.setObject(1, mapping.getTraceId());
-                    statement.setObject(2, mapping.getUserId());
-                    statement.setObject(3, new Timestamp(mapping.getTimestamp()).toLocalDateTime());
-                    statement.addBatch();
-                }
-
-                RetryUtils.retry(statement::executeBatch,
-                                 this::isExceptionRetryable,
-                                 3,
-                                 Duration.ofMillis(100));
-            }
-        }
-
-        protected boolean isExceptionRetryable(Exception e) {
-            return false;
-        }
-    }
-
-    @Override
-    public void write(List<TraceSpan> spans,
-                      List<TraceIdMapping> mappings,
-                      List<TagIndex> tagIndices) {
-
-        TransactionalRunnable runnable = (configuration) -> {
-            List<TraceSpan> summary = spans.stream()
-                                           .filter((span) -> SpanKind.isRootSpan(span.getKind()))
-                                           .collect(Collectors.toList());
-
-            this.writeSpans(summary, Tables.BITHON_TRACE_SPAN_SUMMARY);
-            this.writeSpans(spans, Tables.BITHON_TRACE_SPAN);
-            this.writeMappings(mappings);
-            this.writeTagIndices(tagIndices);
-        };
-        try {
-            if (isTransactionSupported()) {
-                dslContext.transaction(runnable);
-            } else {
-                runnable.run(dslContext.configuration());
-            }
-        } catch (Throwable e) {
-            log.error("Exception when write spans", e);
-        }
-    }
-
-    public static class InsertIndexRunnable implements ConnectionRunnable {
-        private final String insertStatement;
-        private final Collection<Object[]> tagIndices;
-
-        public InsertIndexRunnable(DSLContext dslContext, Collection<Object[]> tagIndices) {
-            this.insertStatement = dslContext.render(dslContext.insertInto(Tables.BITHON_TRACE_SPAN_TAG_INDEX,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.TIMESTAMP,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F1,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F2,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F3,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F4,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F5,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F6,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F7,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F8,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F9,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F10,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F11,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F12,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F13,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F14,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F15,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.F16,
-                                                                           Tables.BITHON_TRACE_SPAN_TAG_INDEX.TRACEID)
-                                                               .values((LocalDateTime) null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null,
-                                                                       null));
-            this.tagIndices = tagIndices;
-        }
-
-        @Override
-        public void run(Connection connection) throws Exception {
-            try (PreparedStatement statement = connection.prepareStatement(insertStatement)) {
-                for (Object[] index : tagIndices) {
-                    for (int i = 0; i < index.length; i++) {
-                        statement.setObject(i + 1, index[i]);
-                    }
-                    statement.addBatch();
-                }
-
-                RetryUtils.retry(statement::executeBatch,
-                                 this::isExceptionRetryable,
-                                 3,
-                                 Duration.ofMillis(100));
-            }
-        }
-
-        protected boolean isExceptionRetryable(Exception e) {
-            return false;
-        }
-    }
-
-    private void writeTagIndices(Collection<TagIndex> tagIndices) {
-        if (CollectionUtils.isEmpty(tagIndices)) {
-            return;
-        }
-
-        Map<String, Object[]> batchValues = new HashMap<>(tagIndices.size());
-        for (TagIndex index : tagIndices) {
-            Object[] values = batchValues.computeIfAbsent(index.getTraceId(), k -> {
-                Object[] r = new Object[18];
-                r[0] = new Timestamp(index.getTimestamp());
-                r[17] = index.getTraceId();
-                return r;
-            });
-
-            int fieldIndex = this.traceStorageConfig.getIndexes().getColumnPos(index.getName());
-            if (fieldIndex == 0 || fieldIndex >= values.length) {
-                // TODO: log error
-                continue;
-            }
-            values[fieldIndex] = index.getValue();
-        }
-
-        dslContext.connection(createInsertIndexRunnable(dslContext, batchValues.values()));
-    }
-
-    protected InsertSpanRunnable createInsertSpanRunnable(Table<?> table, List<TraceSpan> spans) {
-        return new InsertSpanRunnable(dslContext, table, spans) {
-            private final ObjectMapper objectMapper = new ObjectMapper();
-
-            @Override
-            protected Object toTagStore(Map<String, String> tag) {
-                try {
-                    return objectMapper.writeValueAsString(tag);
-                } catch (JsonProcessingException ignored) {
-                    return "{}";
-                }
-            }
-        };
-    }
-
-    protected InsertMappingRunnable createInsertMappingRunnable(DSLContext dslContext, List<TraceIdMapping> mappings) {
-        return new InsertMappingRunnable(dslContext, mappings);
-    }
-
-    protected InsertIndexRunnable createInsertIndexRunnable(DSLContext dslContext, Collection<Object[]> indice) {
-        return new InsertIndexRunnable(dslContext, indice);
-    }
-}
