+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/tracing/reader/NestQueryBuilder.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.jdbc.tracing;
+package org.bithon.server.storage.jdbc.tracing.reader;
 
 import org.bithon.component.commons.expression.IExpression;
 import org.jooq.DSLContext;
